+++ src/containers/Home.js
@@ -56,6 +56,8 @@ const LoadingCover = styled.div`
   }
 `
 
+const CATEGORY = 'category'
+
 const anchors = [
   {
     id: 'latest',
@@ -73,7 +75,7 @@ const anchors = [
     id: 'news-letter',
     label: ''
   }, {
-    id: 'category',
+    id: CATEGORY,
     label: '議題'
   }, {
     id: 'topic',
@@ -197,12 +199,28 @@ class Homepage extends React.Component {
     }
   }
 
+  constructor(props) {
+    super(props)
+    this.categorySection = {}
+  }
+
   componentWillMount() {
     this.props.fetchIndexPageContent()
   }
 
-  componentDidMount() {
-    this.props.fetchCategoriesPostsOnIndexPage()
+  async componentDidMount() {
+    await this.props.fetchCategoriesPostsOnIndexPage()
+    // The following statement implement the mobile header redirection work flow.
+    // Case: user open mobile header slide in menu at article page.
+    // click on category section (議題), the website will direct user to home page
+    // and show the user category section simultaneouslly.
+    if (this.props.location.query.section === CATEGORY && typeof window !== 'undefined' && this.categorySection) {
+      const e = document.getElementById(CATEGORY)
+      if (e) {
+        e.scrollIntoView()
+        this.categorySection.startScrollAnimation()
+      }
+    }
   }
 
   render() {
@@ -251,6 +269,7 @@ class Homepage extends React.Component {
             data={this.props[fieldNames.sections.latestSection]}
             signOutAction={this.props.signOutAction}
             ifAuthenticated={this.props.ifAuthenticated}
+            categoryId={CATEGORY}
           />
           <EditorPicks data={this.props[fieldNames.sections.editorPicksSection]} />
           {latestTopicJSX}
@@ -264,6 +283,7 @@ class Homepage extends React.Component {
           >
             <CategorySection
               data={this.props.categories}
+              ref={(node) => {this.categorySection = node}}
             />
           </Background>
           <Background