+++ src/containers/SignIn.js
@@ -1,15 +1,8 @@
 import React, { Component } from 'react'
-import styled from 'styled-components'
 import { connect } from 'react-redux'
-import { SignInForm, FacebookButton, GoogleButton } from 'twreporter-registration'
+import { SignInForm, FacebookButton, GoogleButton, PageContainer } from 'twreporter-registration'
 import { setHeaderInfo } from '../actions/header'
 import { REGISTRATION, LIGHT } from '../constants/index'
-
-const Container = styled.div`
-  margin: 20px 0;
-  display:flex;
-  justify-content: center;
-`
 
 const TITLE = '登入報導者'
 const REDIRECT_PATH = '/'
@@ -27,17 +20,16 @@ class SignIn extends Component {
     const { params } = this.props
     const redirectPath = params.slug && params.type ? `/${params.type}/${params.slug}` : REDIRECT_PATH
     return (
-      <Container>
+      <PageContainer>
         <SignInForm
           title={TITLE}
           signInRedirectPath={redirectPath}
-          defaultStyle={true}
           {...this.props}
         >
           <FacebookButton />
           <GoogleButton />
         </SignInForm>
-      </Container>
+      </PageContainer>
     )
   }
 }